Refactor http input acceptance test
