+++ tests/acceptance/test_http_input.py
@@ -1,10 +1,5 @@
 # pylint: disable=missing-docstring
 # pylint: disable=line-too-long
-import os
-import re
-import signal
-import subprocess
-import sys
 import time
 from logging import DEBUG, basicConfig, getLogger
 
@@ -12,30 +7,15 @@ import pytest
 import requests
 
 from logprep.util.json_handling import dump_config_as_file
-from tests.acceptance.util import get_default_logprep_config
+from tests.acceptance.util import (
+    get_default_logprep_config,
+    start_logprep,
+    wait_for_output,
+    stop_logprep,
+)
 
 basicConfig(level=DEBUG, format="%(asctime)-15s %(name)-5s %(levelname)-8s: %(message)s")
 logger = getLogger("Logprep-Test")
-
-
-def start_logprep(config_path: str) -> subprocess.Popen:
-    environment = {"PYTHONPATH": "."}
-    return subprocess.Popen(  # nosemgrep
-        f"{sys.executable} logprep/run_logprep.py {config_path}",
-        shell=True,
-        env=environment,
-        stdin=subprocess.PIPE,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-        close_fds=True,
-    )
-
-
-def wait_for_output(proc, expected_output):
-    output = proc.stdout.readline()
-    while expected_output not in output.decode("utf8"):
-        output = proc.stdout.readline()
-        time.sleep(0.1)  # nosemgrep
 
 
 @pytest.fixture(name="config")
@@ -74,14 +54,7 @@ def config_fixture():
 
 
 def teardown_function():
-    # cleanup processes
-    output = subprocess.check_output("ps -x | grep run_logprep", shell=True)  # nosemgrep
-    for line in output.decode("utf8").splitlines():
-        process_id = re.match(r"^\s+(\d+)\s.+", line).group(1)
-        try:
-            os.kill(int(process_id), signal.SIGKILL)
-        except ProcessLookupError:
-            pass
+    stop_logprep()
 
 
 def test_http_input_accepts_message_for_single_pipeline(tmp_path, config):
@@ -91,7 +64,9 @@ def test_http_input_accepts_message_for_single_pipeline(tmp_path, config):
     dump_config_as_file(config_path, config)
     proc = start_logprep(config_path)
     wait_for_output(proc, "Uvicorn running on https://127.0.0.1:9000")
-    requests.post("https://127.0.0.1:9000/plaintext", data="my message", verify=False)  # nosemgrep
+    requests.post(
+        "https://127.0.0.1:9000/plaintext", data="my message", verify=False, timeout=5
+    )  # nosemgrep
     time.sleep(0.5)  # nosemgrep
     assert "my message" in output_path.read_text()
 
@@ -105,10 +80,10 @@ def test_http_input_accepts_message_for_two_pipelines(tmp_path, config):
     proc = start_logprep(config_path)
     wait_for_output(proc, "Uvicorn running on https://127.0.0.1:9001")
     requests.post(  # nosemgrep
-        "https://127.0.0.1:9000/plaintext", data="my first message", verify=False
+        "https://127.0.0.1:9000/plaintext", data="my first message", verify=False, timeout=5
     )
     requests.post(  # nosemgrep
-        "https://127.0.0.1:9001/plaintext", data="my second message", verify=False
+        "https://127.0.0.1:9001/plaintext", data="my second message", verify=False, timeout=5
     )
     time.sleep(0.5)  # nosemgrep
     output_content = output_path.read_text()
@@ -125,13 +100,13 @@ def test_http_input_accepts_message_for_three_pipelines(tmp_path, config):
     proc = start_logprep(config_path)
     wait_for_output(proc, "Uvicorn running on https://127.0.0.1:9002")
     requests.post(  # nosemgrep
-        "https://127.0.0.1:9000/plaintext", data="my first message", verify=False
+        "https://127.0.0.1:9000/plaintext", data="my first message", verify=False, timeout=5
     )
     requests.post(  # nosemgrep
-        "https://127.0.0.1:9001/plaintext", data="my second message", verify=False
+        "https://127.0.0.1:9001/plaintext", data="my second message", verify=False, timeout=5
     )
     requests.post(  # nosemgrep
-        "https://127.0.0.1:9002/plaintext", data="my third message", verify=False
+        "https://127.0.0.1:9002/plaintext", data="my third message", verify=False, timeout=5
     )
     time.sleep(0.5)  # nosemgrep
     output_content = output_path.read_text()
